+++ fabulous/fabric_generator/gds_generator/flows/tile_macro_flow.py
@@ -2,14 +2,13 @@
 
 from decimal import Decimal
 from pathlib import Path
-from typing import Any, Self
+from typing import Self
 
 from librelane.common import GenericDict
-from librelane.config.config import Config
 from librelane.config.variable import Variable
 from librelane.flows.classic import Classic
 from librelane.flows.flow import Flow, FlowException
-from librelane.flows.sequential import SequentialFlow, Substitution, SubstitutionsObject
+from librelane.flows.sequential import SequentialFlow
 from librelane.logging.logger import err, info
 
 from fabulous.fabric_definition.supertile import SuperTile
@@ -25,6 +24,7 @@ from fabulous.fabric_generator.gds_generator.flows.flow_define import (
 from fabulous.fabric_generator.gds_generator.helper import (
     get_pitch,
     get_routing_obstructions,
+    merge_layered_substitutions,
     round_die_area,
 )
 from fabulous.fabric_generator.gds_generator.steps.tile_area_opt import OptMode
@@ -39,44 +39,6 @@ configs = Classic.config_vars + [
         default=False,
     ),
 ]
-
-
-def _merge_layered_substitutions(
-    config_sources: list[Any],
-) -> SubstitutionsObject | None:
-    """Merge `meta.substituting_steps` across layered config sources.
-
-    `Config.load()` overwrites `Config.meta` wholesale for each source in its
-    configs list rather than merging, so a later source with no `meta:` key
-    (e.g. a tile-specific override dict) silently drops an earlier source's
-    `substituting_steps`. This walks the same sources ourselves, in the same
-    order, and concatenates each source's own substitutions so a later
-    source's entries still apply after an earlier source's.
-    """
-    merged: list[tuple[str, Substitution]] = []
-    for source in config_sources:
-        if source is None:
-            continue
-        if isinstance(source, Path | str):
-            if not Path(source).exists():
-                # A missing base/override config file is tolerated elsewhere
-                # in this flow (Config.load treats it as contributing
-                # nothing), so mirror that here instead of letting
-                # Config.get_meta's open() raise on a path that legitimately
-                # may not exist.
-                continue
-            # Config.get_meta only special-cases `str`, not other PathLikes.
-            source = str(source)
-        meta = Config.get_meta(source)
-        if not meta.substituting_steps:
-            continue
-        items = (
-            meta.substituting_steps.items()
-            if isinstance(meta.substituting_steps, dict)
-            else meta.substituting_steps
-        )
-        merged.extend(items)
-    return merged or None
 
 
 class FABulousTileMacroFlow(SequentialFlow):
@@ -113,13 +75,13 @@ class FABulousTileMacroFlow(SequentialFlow):
         """Apply layered `meta.substituting_steps` before construction.
 
         `Config.load()` overwrites its `meta` per config source instead of
-        merging (see `_merge_layered_substitutions`), so `substituting_steps`
+        merging (see `merge_layered_substitutions`), so `substituting_steps`
         from `base_config_path`/`override_config_path` never reaches
         `self.config.meta` by the time `__init__` runs. Resolve substitutions
         from the same layered sources here and apply them by constructing an
         instance of a `.Substitute()`-derived subclass instead.
         """
-        substitutions = _merge_layered_substitutions(
+        substitutions = merge_layered_substitutions(
             [base_config_path, override_config_path, custom_config_overrides]
         )
         target_cls = cls.Substitute(substitutions) if substitutions else cls
